Formatting and creating values for Cell.

+++ src/main/java/com/malyvoj3/csvwvalidator/domain/model/datatypes/DataTypeDefinition.java
@@ -33,7 +33,7 @@ public abstract class DataTypeDefinition {
         throw new UnsupportedOperationException();
     }
 
-    public int getLength() {
+    public Long getLength() {
         throw new UnsupportedOperationException();
     }
